requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-sendgrid-v5==0.8.1
django-simple-email-confirmation==0.70
django-uuslug==1.2.0
future==0.18.2
importlib-metadata==1.6.0
kombu==4.6.8
python-dotenv==0.13.0
python-http-client==3.2.7
python-slugify==4.0.0
pytz==2020.1
redis==3.5.0
sendgrid==6.3.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
vine==1.3.0
zipp==3.1.0
